test(es-index-migrator): dedupe Cassandra schema setup, fix missing enc_payload column
integration_test.go's setupCassandraForRun was missing the enc_payload
column that production's messageColumns SELECT requires, so any query
against it would fail outright with an "undefined column" error. Extract
the schema into a single shared newTestCassandraSession helper used by
both integration_test.go and messagesource_integration_test.go so the two
copies can no longer drift.
