+++ data-migration/es-index-migrator/integration_test.go
@@ -46,64 +46,15 @@ func getESDoc(t *testing.T, esURL, index, docID string) map[string]any {
 	return result.Source
 }
 
-// setupCassandraForRun creates an isolated keyspace (via
-// testutil.CassandraKeyspace) with the Participant/Card UDTs and the
-// messages_by_room table this service reads — mirroring
-// messagesource_integration_test.go's setupCassandra — and returns the
-// keyspace name and host so the caller can build a config for run() to
-// dial its own session against, plus a keyspace-pinned session for seeding
-// test rows (production queries are unqualified table names, so the seed
-// insert needs the same pinning).
+// setupCassandraForRun returns the keyspace name and host so the caller can
+// build a config for run() to dial its own session against, plus a
+// keyspace-pinned session for seeding test rows (production queries use
+// unqualified table names, so the seed insert needs the same pinning). Schema
+// setup is shared with messagesource_integration_test.go's setupCassandra via
+// newTestCassandraSession in cassandra_test_helpers_test.go.
 func setupCassandraForRun(t *testing.T, prefix string) (keyspace, hostAddr string, pinned *gocql.Session) {
 	t.Helper()
-	keyspace, adminSession, hostAddr := testutil.CassandraKeyspace(t, prefix)
-	cql := func(format string) string { return fmt.Sprintf(format, keyspace) }
-
-	stmts := []string{
-		cql(`CREATE TYPE IF NOT EXISTS %s."Participant" (
-			id           TEXT,
-			eng_name     TEXT,
-			company_name TEXT,
-			app_id       TEXT,
-			app_name     TEXT,
-			is_bot       BOOLEAN,
-			account      TEXT
-		)`),
-		cql(`CREATE TYPE IF NOT EXISTS %s."Card" (
-			template TEXT,
-			data     BLOB
-		)`),
-		cql(`CREATE TABLE IF NOT EXISTS %s.messages_by_room (
-			room_id                  TEXT,
-			bucket                   BIGINT,
-			created_at               TIMESTAMP,
-			message_id               TEXT,
-			sender                   FROZEN<"Participant">,
-			msg                      TEXT,
-			attachments              LIST<BLOB>,
-			card                     FROZEN<"Card">,
-			tshow                    BOOLEAN,
-			thread_parent_id         TEXT,
-			thread_parent_created_at TIMESTAMP,
-			deleted                  BOOLEAN,
-			site_id                  TEXT,
-			edited_at                TIMESTAMP,
-			updated_at               TIMESTAMP,
-			PRIMARY KEY ((room_id, bucket), created_at, message_id)
-		) WITH CLUSTERING ORDER BY (created_at DESC, message_id DESC)`),
-	}
-	for _, stmt := range stmts {
-		require.NoError(t, adminSession.Query(stmt).Exec())
-	}
-
-	cluster := gocql.NewCluster(hostAddr)
-	cluster.Consistency = gocql.One
-	cluster.DisableInitialHostLookup = true
-	cluster.Keyspace = keyspace
-	ksSession, err := cluster.CreateSession()
-	require.NoError(t, err)
-	t.Cleanup(func() { ksSession.Close() })
-	return keyspace, hostAddr, ksSession
+	return newTestCassandraSession(t, prefix)
 }
 
 func TestRun_EndToEndBackfillIsIdempotentOnRerun(t *testing.T) {
